move metadata around
