+++ hydromt/_typing/metadata.py
