+++ hydromt/drivers/dataframe/pandas_driver.py
@@ -6,10 +6,15 @@ from typing import List, Optional
 
 import pandas as pd
 
-from hydromt._typing import NoDataStrategy, StrPath, TimeRange, Variables
+from hydromt._typing import (
+    NoDataStrategy,
+    SourceMetadata,
+    StrPath,
+    TimeRange,
+    Variables,
+)
 from hydromt._utils.unused_kwargs import warn_on_unused_kwargs
 from hydromt.drivers.dataframe import DataFrameDriver
-from hydromt.metadata import SourceMetadata
 
 logger: Logger = getLogger(__name__)
