+++ hydromt/drivers/geodataframe/pyogrio_driver.py
@@ -7,11 +7,10 @@ import geopandas as gpd
 from pyogrio import read_dataframe, read_info, write_dataframe
 from pyproj import CRS
 
-from hydromt._typing import Bbox, Geom, StrPath
+from hydromt._typing import Bbox, Geom, SourceMetadata, StrPath
 from hydromt._typing.error import NoDataStrategy
 from hydromt._utils.unused_kwargs import warn_on_unused_kwargs
 from hydromt.drivers.geodataframe.geodataframe_driver import GeoDataFrameDriver
-from hydromt.metadata import SourceMetadata
 
 logger: Logger = getLogger(__name__)
