+++ hydromt/drivers/raster/raster_xarray_driver.py
@@ -8,12 +8,18 @@ from typing import Callable, List, Optional
 
 import xarray as xr
 
-from hydromt._typing import Geom, StrPath, TimeRange, Variables, ZoomLevel
+from hydromt._typing import (
+    Geom,
+    SourceMetadata,
+    StrPath,
+    TimeRange,
+    Variables,
+    ZoomLevel,
+)
 from hydromt._typing.error import NoDataStrategy
 from hydromt._utils.unused_kwargs import warn_on_unused_kwargs
 from hydromt.drivers.preprocessing import PREPROCESSORS
 from hydromt.drivers.raster.rasterdataset_driver import RasterDatasetDriver
-from hydromt.metadata import SourceMetadata
 
 logger: Logger = getLogger(__name__)
