+++ tests/data_adapter/test_data_adapter.py
@@ -606,7 +606,8 @@ def test_dataframe_time(df_time, tmpdir):
 
 
 @pytest.mark.skip(reason="Needs implementation of all raster Drivers.")
-# FIXME: Partially tested in "test_rasterio_driver", just missing the cache=True option.
+# TODO: Partially tested in "test_rasterio_driver", just missing the cache=True option.
+# https://github.com/Deltares/hydromt/issues/897
 def test_cache_vrt(tmpdir, rioda_large):
     # write vrt data
     name = "tiled"
