+++ tests/data_adapter/test_dataframe_adapter.py
@@ -2,8 +2,8 @@ from copy import copy
 
 import pandas as pd
 
+from hydromt._typing import SourceMetadata
 from hydromt.data_adapter.dataframe import DataFrameAdapter
-from hydromt.metadata import SourceMetadata
 
 
 class TestDataFrameAdapter:
