+++ tests/data_adapter/test_raster_dataset_adapter.py
@@ -2,9 +2,8 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from hydromt._typing import NoDataException, NoDataStrategy
+from hydromt._typing import NoDataException, NoDataStrategy, SourceMetadata
 from hydromt.data_adapter.rasterdataset import RasterDatasetAdapter
-from hydromt.metadata import SourceMetadata
 
 
 class TestRasterDatasetAdapter:
