+++ tests/data_sources/test_geo_dataframe_source.py
@@ -12,13 +12,12 @@ from pystac import Asset as StacAsset
 from pystac import Catalog as StacCatalog
 from pystac import Item as StacItem
 
-from hydromt._typing import StrPath
+from hydromt._typing import SourceMetadata, StrPath
 from hydromt._typing.error import ErrorHandleMethod
 from hydromt.data_adapter.geodataframe import GeoDataFrameAdapter
 from hydromt.data_catalog import DataCatalog
 from hydromt.data_source.geodataframe import GeoDataFrameSource
 from hydromt.drivers import GeoDataFrameDriver, PyogrioDriver
-from hydromt.metadata import SourceMetadata
 from hydromt.metadata_resolver.convention_resolver import ConventionResolver
