+++ tests/data_sources/test_geo_dataset_source.py
@@ -5,11 +5,10 @@ import pytest
 import xarray as xr
 from pydantic import ValidationError
 
-from hydromt._typing import StrPath
+from hydromt._typing import SourceMetadata, StrPath
 from hydromt.data_adapter import GeoDatasetAdapter
 from hydromt.data_source import GeoDatasetSource
 from hydromt.drivers import GeoDatasetDriver
-from hydromt.metadata import SourceMetadata
 
 
 @pytest.fixture()
